+++ src/entities/BasicEnemy.java
@@ -32,19 +32,19 @@ public class BasicEnemy extends GameObject {
 		if (y <= 0 || y >= Game.HEIGHT - 32) velY *= -1;
 		if (x <= 0 || x >= Game.WIDTH - 16) velX *= -1;
 		
-		handler.addObject(new  Trail(x, y, ID.Trail, Color.RED, 16, 16, 0.02f, handler));
+		handler.addObject(new  Trail((int)x, (int)y, ID.Trail, Color.RED, 16, 16, 0.02f, handler));
 		
 	}
 
 	@Override
 	public void render(Graphics g) {
 		g.setColor(Color.RED);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 		
 	}
 	
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 
 }
